chore: rename constants
