+++ landing/pages/connacol-page/src/connacol.constants.ts
