+++ landing/pages/contract-offer-page/src/contract-offer.constants.ts
