+++ landing/pages/feeling-of-time-page/src/feeling-of-time.constants.ts
