+++ landing/pages/index-page/src/index-page.constants.ts
