+++ landing/pages/opening-the-rhythm-page/src/opening-the-rhythm.constants.ts
