+++ landing/pages/seventh-heaven-page/src/seventh-heaven.constants.ts
